:truck: Refactor #31: remote 계층 model 폴더 정리
