+++ app/src/main/java/com/umc/edison/remote/datasources/LabelRemoteDataSourceImpl.kt
@@ -4,7 +4,7 @@ import com.umc.edison.data.datasources.LabelRemoteDataSource
 import com.umc.edison.data.model.LabelEntity
 import com.umc.edison.remote.api.BubbleSpaceApiService
 import com.umc.edison.remote.api.SyncApiService
-import com.umc.edison.remote.model.toSyncLabelRequest
+import com.umc.edison.remote.model.sync.toSyncLabelRequest
 import javax.inject.Inject
 
 class LabelRemoteDataSourceImpl @Inject constructor(
