+++ app/src/main/java/com/umc/edison/remote/model/bubble_space/GetLabelDetailResponse.kt
@@ -1,9 +1,10 @@
-package com.umc.edison.remote.model
+package com.umc.edison.remote.model.bubble_space
 
 import androidx.compose.ui.graphics.Color
 import com.google.gson.annotations.SerializedName
 import com.umc.edison.data.model.BubbleEntity
 import com.umc.edison.data.model.LabelEntity
+import com.umc.edison.remote.model.RemoteMapper
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
