+++ app/src/main/java/com/umc/edison/remote/model/sync/SyncLabelResponse.kt
@@ -1,8 +1,10 @@
-package com.umc.edison.remote.model
+package com.umc.edison.remote.model.sync
 
 import androidx.compose.ui.graphics.Color
 import com.google.gson.annotations.SerializedName
 import com.umc.edison.data.model.LabelEntity
+import com.umc.edison.remote.model.RemoteMapper
+import com.umc.edison.remote.model.toDate
 
 data class SyncLabelResponse(
     @SerializedName("labelId") val id: Int,
